+++ api/package.json
@@ -25,27 +25,26 @@
   },
   "license": "GPL-3.0",
   "peerDependencies": {
-    "@polkadot/api": "^8.0.1",
+    "@polkadot/api": "^8.0.2",
     "rxjs": "^7.5.5"
   },
   "devDependencies": {
     "@babel/preset-env": "7.16.11",
     "@babel/preset-typescript": "7.16.7",
+    "@rollup/plugin-commonjs": "21.0.3",
+    "@rollup/plugin-json": "4.1.0",
+    "@rollup/plugin-node-resolve": "13.2.0",
     "@types/jest": "27.4.1",
+    "@typescript-eslint/eslint-plugin": "5.18.0",
+    "@typescript-eslint/parser": "5.18.0",
     "babel-cli": "6.26.0",
     "babel-core": "6.26.3",
     "babel-jest": "27.5.1",
-    "jest": "27.5.1",
-    "tslint": "6.1.3",
-    "@rollup/plugin-commonjs": "21.0.3",
-    "@rollup/plugin-json": "4.1.0",
-    "@rollup/plugin-node-resolve": "13.1.3",
-    "@rollup/plugin-typescript": "8.3.1",
-    "@typescript-eslint/eslint-plugin": "5.18.0",
-    "@typescript-eslint/parser": "5.18.0",
     "eslint": "8.12.0",
+    "jest": "27.5.1",
     "prettier": "2.6.2",
     "rollup": "2.70.1",
+    "tslint": "6.1.3",
     "typescript": "4.6.3"
   },
   "directories": {
@@ -63,5 +62,10 @@
   "bugs": {
     "url": "https://github.com/gear-tech/gear-js/issues"
   },
-  "homepage": "https://github.com/gear-tech/gear-js/tree/master/api#readme"
+  "homepage": "https://github.com/gear-tech/gear-js/tree/master/api#readme",
+  "dependencies": {
+    "@types/rollup-plugin-peer-deps-external": "^2.2.1",
+    "rollup-plugin-peer-deps-external": "^2.2.4",
+    "rollup-plugin-typescript2": "^0.31.2"
+  }
 }
