+++ api/src/GasSpent.ts
@@ -1,8 +1,8 @@
-import { AnyJson } from '@polkadot/types-codec/types';
-import { isHex, isString } from '@polkadot/util';
+import { isHex, isString, isU8a } from '@polkadot/util';
 import { u64 } from '@polkadot/types';
 import { CreateType } from './create-type';
-import { Hex, Metadata } from './interfaces';
+import { Metadata } from './types/interfaces';
+import { Hex, PayloadType } from './types';
 import { GearApi } from './GearApi';
 import { createPayload } from './utils';
 import { GetGasSpentError } from './errors/program.errors';
@@ -15,8 +15,12 @@ export class GearGasSpent {
     this.createType = new CreateType(this.api);
   }
 
-  private getPayload(payload: Hex | AnyJson, metaOrTypeOfPayload: string | Metadata, meta_type: string): Hex {
-    if (isHex(payload)) {
+  private getPayload(
+    payload: PayloadType,
+    metaOrTypeOfPayload: string | Metadata,
+    meta_type: string,
+  ): Hex | Uint8Array {
+    if (isHex(payload) || isU8a(payload)) {
       return payload;
     }
     if (!metaOrTypeOfPayload) {
@@ -56,7 +60,7 @@ export class GearGasSpent {
   async init(
     sourceId: Hex,
     code: Hex | Buffer,
-    payload: Hex | AnyJson,
+    payload: PayloadType,
     value: number | string,
     meta?: Metadata,
   ): Promise<u64>;
@@ -84,7 +88,7 @@ export class GearGasSpent {
   async init(
     sourceId: Hex,
     code: Hex | Buffer,
-    payload: Hex | AnyJson,
+    payload: PayloadType,
     value: number | string,
     typeOfPayload?: string,
   ): Promise<u64>;
@@ -101,7 +105,7 @@ export class GearGasSpent {
   async init(
     sourceId: Hex,
     code: Hex | Buffer,
-    payload: Hex | AnyJson,
+    payload: PayloadType,
     value: number | string,
     metaOrTypeOfPayload?: string | Metadata,
   ): Promise<u64>;
@@ -109,7 +113,7 @@ export class GearGasSpent {
   async init(
     sourceId: Hex,
     code: Hex | Buffer,
-    payload: Hex | AnyJson,
+    payload: PayloadType,
     value: number | string,
     metaOrTypeOfPayload?: string | Metadata,
   ): Promise<u64> {
@@ -151,7 +155,7 @@ export class GearGasSpent {
   async handle(
     sourceId: Hex,
     destinationId: Hex | Buffer,
-    payload: Hex | AnyJson,
+    payload: PayloadType,
     value: number | string,
     meta?: Metadata,
   ): Promise<u64>;
@@ -180,7 +184,7 @@ export class GearGasSpent {
   async handle(
     sourceId: Hex,
     destinationId: Hex | Buffer,
-    payload: Hex | AnyJson,
+    payload: PayloadType,
     value: number | string,
     typeOfPayload?: string,
   ): Promise<u64>;
@@ -197,7 +201,7 @@ export class GearGasSpent {
   async handle(
     sourceId: Hex,
     destinationId: Hex | Buffer,
-    payload: Hex | AnyJson,
+    payload: PayloadType,
     value: number | string,
     metaOrTypeOfPayload?: string | Metadata,
   ): Promise<u64>;
@@ -205,7 +209,7 @@ export class GearGasSpent {
   async handle(
     sourceId: Hex,
     destinationId: Hex,
-    payload: Hex | AnyJson,
+    payload: PayloadType,
     value: number | string,
     metaOrTypeOfPayload?: string | Metadata,
   ): Promise<u64> {
@@ -245,7 +249,7 @@ export class GearGasSpent {
     sourceId: Hex,
     messageId: Hex,
     exitCode: number,
-    payload: Hex | AnyJson,
+    payload: PayloadType,
     value: number | string,
     meta?: Metadata,
   ): Promise<u64>;
@@ -278,7 +282,7 @@ export class GearGasSpent {
     sourceId: Hex,
     messageId: Hex,
     exitCode: number,
-    payload: Hex | AnyJson,
+    payload: PayloadType,
     value: number | string,
     typeOfPayload?: string,
   ): Promise<u64>;
@@ -297,7 +301,7 @@ export class GearGasSpent {
     sourceId: Hex,
     messageId: Hex,
     exitCode: number,
-    payload: Hex | AnyJson,
+    payload: PayloadType,
     value: number | string,
     metaOrTypeOfPayload?: string | Metadata,
   ): Promise<u64>;
@@ -306,7 +310,7 @@ export class GearGasSpent {
     sourceId: Hex,
     messageId: Hex,
     exitCode: number,
-    payload: Hex | AnyJson,
+    payload: PayloadType,
     value: number | string,
     metaOrTypeOfPayload?: string | Metadata,
   ): Promise<u64> {
