+++ api/src/events-types/index.ts
@@ -1,0 +1,2 @@
+export * from './Event';
+export * from './EventData';
