+++ api/src/index.ts
@@ -13,9 +13,11 @@ export { GearBlock } from './Blocks';
 export { GearGasSpent } from './GasSpent';
 export { GearStorage } from './Storage';
 export { GearClaimValue } from './Claim';
+export * from './Transaction';
 export * from './wasm';
 export * from './CreateType';
 export * from './create-type';
-export * from './interfaces';
-export * from './types';
+export * as interfaces from './types/interfaces';
+export * as types from './types';
+export * from './events';
 export * from './utils';
