--- api/src/interfaces/mailbox.ts
@@ -1,4 +1,0 @@
-import { AccountId, MessageId } from './gear-api';
-import { HumanedMessage } from './message';
-
-export type IMailbox = [[AccountId, MessageId], HumanedMessage][];
