--- api/src/interfaces/transaction.ts
@@ -1,3 +1,0 @@
-import { ISubmittableResult } from '@polkadot/types/types';
-
-export type TransactionStatusCb = (result: ISubmittableResult, extra: undefined) => void | Promise<void>;
