+++ api/src/types/common.ts
@@ -1,0 +1,5 @@
+import { i32 } from '@polkadot/types';
+
+export declare type Hex = `0x${string}`;
+
+export declare type ExitCode = i32;
