+++ api/src/types/ids.ts
@@ -1,0 +1,8 @@
+import { Hex } from './common';
+import { H256 } from '@polkadot/types/interfaces';
+
+export declare type ProgramId = Hex | H256 | string;
+
+export declare type MessageId = Hex;
+
+export declare type AccountId = string | Hex;
