+++ api/src/types/index.ts
@@ -1,3 +1,5 @@
-export * from './Event';
-export * from './EventData';
-export * from './Transaction';
+export * from './common';
+export * from './payload';
+export * from './ids';
+export * from './maibox';
+export * from './transaction';
