+++ api/src/types/interfaces/debug.ts
