+++ api/src/types/interfaces/gear-type.ts
@@ -1,5 +1,3 @@
 export interface GearType {
   [index: string]: any;
 }
-
-export declare type Hex = `0x${string}`;
