+++ api/src/types/interfaces/index.ts
@@ -5,9 +5,8 @@ export * from './metadata';
 export * from './debug';
 export * from './keyring';
 export * from './program';
-export * from './transaction';
 export * from './events';
 export * from './system';
 export * from './augment-api-rpc';
 export * from './augment-api-tx';
-export * from './mailbox';
+export * from './system';
