+++ api/src/types/interfaces/keyring.ts
@@ -1,4 +1,5 @@
 import { KeyringPair$Json } from '@polkadot/keyring/types';
+
 export interface CreateKeyring {
   keyPairJson?: KeyringPair$Json | string;
   seed?: Uint8Array | string;
