+++ api/src/types/interfaces/program.ts
@@ -1,5 +1,6 @@
-import { Vec, Bytes, u8, u32, u64, Null, Map, Type, Compact } from '@polkadot/types';
+import { Vec, Bytes, u8, u32, u64, Null, Map, Type } from '@polkadot/types';
 import { H256 } from '@polkadot/types/interfaces';
+
 export declare interface Reason extends Bytes {
   isError: Boolean;
   asError: Null;
