+++ api/src/types/maibox.ts
@@ -1,0 +1,4 @@
+import { AccountId, MessageId } from './ids';
+import { HumanedMessage } from './interfaces';
+
+export type MailboxType = [[AccountId, MessageId], HumanedMessage][];
