+++ api/src/types/payload.ts
@@ -1,0 +1,5 @@
+import { Hex } from './common';
+import { Bytes } from '@polkadot/types';
+import { AnyJson } from '@polkadot/types/types';
+
+export type PayloadType = Hex | Uint8Array | Bytes | AnyJson;
